feat: 각각의 자동차 독립성을 위해 Car 클래스 추가
